requirements.txt
torch
transformers
pandas
numpy
scikit-learn
nltk
tqdm
pyyaml
matplotlib
seaborn
streamlit
torch
transformers
pillow
pandas
numpy
plotly
streamlit-lottie
tqdm
scikit-learn